Basic QSPI support on STM32
When QSPI is used, only half-duplex single line communication is supported and modes are limited to 0 and 3.

If io2 and io3 pins are set, they will be configured but will be fixed at 0 and 1 respectively and not otherwise used (this may be useful when flash chip is attached directly and needs to have HOLD and WP pins in definite state).

CL: spi: Basic QSPI support on STM32

PUBLISHED_FROM=0563a2c4d2b4628043541cca878254b6cc8527eb